Move nonemin, nonemax into _utils/reduction
